# More efficient approach to checking /v2/editions pagination